refactor(credit-cards): split CreditCardsPage into smaller components
Extracts CreditCardsSummary (KPI cards), CreditCardsGrid (card list)
and CreditCardMovementsTable (filters + movements table) out of
index.tsx, addressing the react-doctor no-giant-component warning
(index.tsx goes from 375 to 189 lines).
